ATry alternate autocreate method.
