+++ client/src/components/Header/Header.tsx
@@ -4,15 +4,10 @@ import logo from 'images/inferno_logo.png';
 import { Box, Button, Container } from '@material-ui/core';
 import { useHistory } from 'react-router-dom';
 
-export interface HeaderProps {
-  setTestSuiteChosen: (id: string) => void;
-}
-
-const Header: FC<HeaderProps> = ({ setTestSuiteChosen }) => {
+const Header: FC<unknown> = () => {
   const styles = useStyles();
   const history = useHistory();
   function returnHome() {
-    setTestSuiteChosen('');
     history.push('/');
   }
   return (
